XWindow: harmonize Doxygen style and add an XDisplay member
Use the harmonized Doxygen style as used in libcosmos now.

Also introduce a permanent XDisplay member variable in XWindow.
Currently a lot of XDisplay::getInstance() calls are used that are
potentially less efficient and also more verbose.

This way the XWindow class might also be able to be more easily used
with a non-default XDisplay in the future.