global: harmonize naming scheme, rename X++ -> xpp
The plusses in the name give us trouble e.g. on GitHub where they cannot
be used in repository names. Furthermore other libraries exist on Linux
that have similar names, one is even libx++ lower-case from the xplane
game.

Thus rename the library and include folder to xpp lower case. This
matches the namespace xpp:: used in the code and thus all identifiers
should be harmonized.

With this also bump the SO version to 0.2.0.